+++ kernel-rs/src/exec.rs
@@ -14,17 +14,17 @@ use crate::{
 use core::ptr;
 
 pub unsafe fn exec(mut path: *mut libc::c_char, mut argv: *mut *mut libc::c_char) -> i32 {
-    let mut oldsz: u64 = 0;
-    let mut current_block: u64;
+    let mut oldsz: usize = 0;
+    let mut current_block: usize;
     let mut s: *mut libc::c_char = ptr::null_mut();
     let mut last: *mut libc::c_char = ptr::null_mut();
     let mut i: i32 = 0;
     let mut off: i32 = 0;
-    let mut argc: u64 = 0;
-    let mut sz: u64 = 0;
-    let mut sp: u64 = 0;
-    let mut ustack: [u64; 33] = [0; 33];
-    let mut stackbase: u64 = 0;
+    let mut argc: usize = 0;
+    let mut sz: usize = 0;
+    let mut sp: usize = 0;
+    let mut ustack: [usize; 33] = [0; 33];
+    let mut stackbase: usize = 0;
     let mut elf: ElfHdr = Default::default();
     let mut ip: *mut Inode = ptr::null_mut();
     let mut ph: ProgHdr = Default::default();
@@ -43,17 +43,17 @@ pub unsafe fn exec(mut path: *mut libc::c_char, mut argv: *mut *mut libc::c_char
     if readi(
         ip,
         0 as i32,
-        &mut elf as *mut ElfHdr as u64,
+        &mut elf as *mut ElfHdr as usize,
         0 as i32 as u32,
-        ::core::mem::size_of::<ElfHdr>() as u64 as u32,
-    ) as u64
-        == ::core::mem::size_of::<ElfHdr>() as u64
+        ::core::mem::size_of::<ElfHdr>() as usize as u32,
+    ) as usize
+        == ::core::mem::size_of::<ElfHdr>() as usize
         && elf.magic == ELF_MAGIC
     {
         pagetable = proc_pagetable(p);
         if !pagetable.is_null() {
             // Load program into memory.
-            sz = 0 as i32 as u64;
+            sz = 0 as i32 as usize;
             i = 0 as i32;
             off = elf.phoff as i32;
             loop {
@@ -64,11 +64,11 @@ pub unsafe fn exec(mut path: *mut libc::c_char, mut argv: *mut *mut libc::c_char
                 if readi(
                     ip,
                     0 as i32,
-                    &mut ph as *mut ProgHdr as u64,
+                    &mut ph as *mut ProgHdr as usize,
                     off as u32,
-                    ::core::mem::size_of::<ProgHdr>() as u64 as u32,
-                ) as u64
-                    != ::core::mem::size_of::<ProgHdr>() as u64
+                    ::core::mem::size_of::<ProgHdr>() as usize as u32,
+                ) as usize
+                    != ::core::mem::size_of::<ProgHdr>() as usize
                 {
                     current_block = 7080392026674647309;
                     break;
@@ -83,11 +83,11 @@ pub unsafe fn exec(mut path: *mut libc::c_char, mut argv: *mut *mut libc::c_char
                         break;
                     }
                     sz = uvmalloc(pagetable, sz, ph.vaddr.wrapping_add(ph.memsz));
-                    if sz == 0 as i32 as u64 {
+                    if sz == 0 as i32 as usize {
                         current_block = 7080392026674647309;
                         break;
                     }
-                    if ph.vaddr.wrapping_rem(PGSIZE as u64) != 0 as i32 as u64 {
+                    if ph.vaddr.wrapping_rem(PGSIZE as usize) != 0 as i32 as usize {
                         current_block = 7080392026674647309;
                         break;
                     }
@@ -98,7 +98,7 @@ pub unsafe fn exec(mut path: *mut libc::c_char, mut argv: *mut *mut libc::c_char
                     }
                 }
                 i += 1;
-                off = (off as u64).wrapping_add(::core::mem::size_of::<ProgHdr>() as u64) as i32
+                off = (off as usize).wrapping_add(::core::mem::size_of::<ProgHdr>() as usize) as i32
                     as i32
             }
             match current_block {
@@ -112,32 +112,34 @@ pub unsafe fn exec(mut path: *mut libc::c_char, mut argv: *mut *mut libc::c_char
 
                     // Allocate two pages at the next page boundary.
                     // Use the second as the user stack.
-                    sz = sz.wrapping_add(PGSIZE as u64).wrapping_sub(1 as i32 as u64)
-                        & !(PGSIZE - 1 as i32) as u64;
-                    sz = uvmalloc(pagetable, sz, sz.wrapping_add((2 as i32 * PGSIZE) as u64));
-                    if sz != 0 as i32 as u64 {
-                        uvmclear(pagetable, sz.wrapping_sub((2 as i32 * PGSIZE) as u64));
+                    sz = sz
+                        .wrapping_add(PGSIZE as usize)
+                        .wrapping_sub(1 as i32 as usize)
+                        & !(PGSIZE - 1 as i32) as usize;
+                    sz = uvmalloc(pagetable, sz, sz.wrapping_add((2 as i32 * PGSIZE) as usize));
+                    if sz != 0 as i32 as usize {
+                        uvmclear(pagetable, sz.wrapping_sub((2 as i32 * PGSIZE) as usize));
                         sp = sz;
-                        stackbase = sp.wrapping_sub(PGSIZE as u64);
+                        stackbase = sp.wrapping_sub(PGSIZE as usize);
 
                         // Push argument strings, prepare rest of stack in ustack.
-                        argc = 0 as i32 as u64;
+                        argc = 0 as i32 as usize;
                         loop {
                             if (*argv.offset(argc as isize)).is_null() {
                                 current_block = 4567019141635105728;
                                 break;
                             }
-                            if argc >= MAXARG as u64 {
+                            if argc >= MAXARG as usize {
                                 current_block = 7080392026674647309;
                                 break;
                             }
-                            sp = (sp as u64).wrapping_sub(
-                                (strlen(*argv.offset(argc as isize)) + 1 as i32) as u64,
-                            ) as u64 as u64;
+                            sp = (sp as usize).wrapping_sub(
+                                (strlen(*argv.offset(argc as isize)) + 1 as i32) as usize,
+                            ) as usize as usize;
 
                             // riscv sp must be 16-byte aligned
-                            sp = (sp as u64).wrapping_sub(sp.wrapping_rem(16 as i32 as u64)) as u64
-                                as u64;
+                            sp = (sp as usize).wrapping_sub(sp.wrapping_rem(16 as i32 as usize))
+                                as usize as usize;
                             if sp < stackbase {
                                 current_block = 7080392026674647309;
                                 break;
@@ -146,7 +148,7 @@ pub unsafe fn exec(mut path: *mut libc::c_char, mut argv: *mut *mut libc::c_char
                                 pagetable,
                                 sp,
                                 *argv.offset(argc as isize),
-                                (strlen(*argv.offset(argc as isize)) + 1 as i32) as u64,
+                                (strlen(*argv.offset(argc as isize)) + 1 as i32) as usize,
                             ) < 0 as i32
                             {
                                 current_block = 7080392026674647309;
@@ -158,22 +160,22 @@ pub unsafe fn exec(mut path: *mut libc::c_char, mut argv: *mut *mut libc::c_char
                         match current_block {
                             7080392026674647309 => {}
                             _ => {
-                                ustack[argc as usize] = 0 as i32 as u64;
+                                ustack[argc as usize] = 0 as i32 as usize;
 
                                 // push the array of argv[] pointers.
-                                sp = (sp as u64).wrapping_sub(
-                                    argc.wrapping_add(1 as i32 as u64)
-                                        .wrapping_mul(::core::mem::size_of::<u64>() as u64),
-                                ) as u64 as u64;
-                                sp = (sp as u64).wrapping_sub(sp.wrapping_rem(16 as i32 as u64))
-                                    as u64 as u64;
+                                sp = (sp as usize).wrapping_sub(
+                                    argc.wrapping_add(1 as i32 as usize)
+                                        .wrapping_mul(::core::mem::size_of::<usize>() as usize),
+                                ) as usize as usize;
+                                sp = (sp as usize).wrapping_sub(sp.wrapping_rem(16 as i32 as usize))
+                                    as usize as usize;
                                 if sp >= stackbase
                                     && copyout(
                                         pagetable,
                                         sp,
                                         ustack.as_mut_ptr() as *mut libc::c_char,
-                                        argc.wrapping_add(1 as i32 as u64)
-                                            .wrapping_mul(::core::mem::size_of::<u64>() as u64),
+                                        argc.wrapping_add(1 as i32 as usize)
+                                            .wrapping_mul(::core::mem::size_of::<usize>() as usize),
                                     ) >= 0 as i32
                                 {
                                     // arguments to user main(argc, argv)
@@ -193,7 +195,8 @@ pub unsafe fn exec(mut path: *mut libc::c_char, mut argv: *mut *mut libc::c_char
                                     safestrcpy(
                                         (*p).name.as_mut_ptr(),
                                         last,
-                                        ::core::mem::size_of::<[libc::c_char; 16]>() as u64 as i32,
+                                        ::core::mem::size_of::<[libc::c_char; 16]>() as usize
+                                            as i32,
                                     );
 
                                     // Commit to the user image.
@@ -234,21 +237,21 @@ pub unsafe fn exec(mut path: *mut libc::c_char, mut argv: *mut *mut libc::c_char
 /// Returns 0 on success, -1 on failure.
 unsafe fn loadseg(
     mut pagetable: pagetable_t,
-    mut va: u64,
+    mut va: usize,
     mut ip: *mut Inode,
     mut offset: u32,
     mut sz: u32,
 ) -> i32 {
     let mut i: u32 = 0;
-    if va.wrapping_rem(PGSIZE as u64) != 0 as i32 as u64 {
+    if va.wrapping_rem(PGSIZE as usize) != 0 as i32 as usize {
         panic(
             b"loadseg: va must be page aligned\x00" as *const u8 as *const libc::c_char
                 as *mut libc::c_char,
         );
     }
     while i < sz {
-        let pa = walkaddr(pagetable, va.wrapping_add(i as u64));
-        if pa == 0 as i32 as u64 {
+        let pa = walkaddr(pagetable, va.wrapping_add(i as usize));
+        if pa == 0 as i32 as usize {
             panic(
                 b"loadseg: address should exist\x00" as *const u8 as *const libc::c_char
                     as *mut libc::c_char,
